Refactor: Refactor support actions
